requirements.txt
openai
fastapi
uvicorn
python-dotenv
jinja2